migrating to postgres

+++ app/controllers/article_controller.rb
@@ -1,7 +1,6 @@
 # -*- encoding : utf-8 -*-
 class ArticleController < ApplicationController
   before_filter :authenticate_user!, :except => [:show]
-  before_filter :check_permissions, :except => [:show, :new, :create]
 
   def new
     @article = Article.new
@@ -10,7 +9,6 @@ class ArticleController < ApplicationController
   end
 
   def create
-    @article = Article.new(params[:article])
     @post = Post.new({:author_id=>current_user[:id],
                       :title => params[:post][:title],
                       :status => params[:post][:status],
@@ -18,51 +16,57 @@ class ArticleController < ApplicationController
                       :tournament_id => params[:post][:tournament_id]
                      })
 
-    @post.resource = @article
+    @post.resource = Article.new(params[:article])
+
+    authorize @post, :update?
+
     respond_to do |format|
       if @post.save
         flash[:notice] = "Статья успешно сохранена"
         format.html { redirect_to(:controller => 'blog', :action => 'index') }
-        format.xml  { render :xml => @post, :status => :created, :location => @post }
       else
         format.html { render :action => "new" }
-        format.xml  { render :xml => @article.errors, :status => :unprocessable_entity }
       end
     end
   end
 
   def edit
     @article = Article.find(params[:id])
+
+    authorize @article.post, :update?
+
     @article_image = @article.article_image
     render :layout => "app_without_sidebar"
   end
 
   def update
     @article = Article.find(params[:id])
+
+    authorize @article.post, :update?
+
     respond_to do |format|
 
       if @article.update_attributes(params[:article]) && @article.post.update_attributes(params[:post])
         flash[:notice] = "Статья успешно сохранена"
         format.html { redirect_to(:controller => 'blog', :action => 'index') }
-        format.xml  { head :ok }
       else
         format.html { render :action => "edit" }
-        format.xml  { render :xml => @article.errors, :status => :unprocessable_entity }
       end
     end
   end
 
   def delete
       @article = Article.find(params[:id])
+
+      authorize @article.post, :update?
+
       respond_to do |format|
         if user_signed_in? && current_user[:id]==1
           if @article.post.resource.destroy
             flash[:notice] = "Статья успешно удалена"
             format.html { redirect_to(:controller => 'blog', :action => 'index') }
-            format.xml  { head :ok }
           else
             format.html { render :action => "edit", :id => params[:id] }
-            format.xml  { render :xml => @article.errors, :status => :unprocessable_entity }
           end
         end
       end
@@ -70,17 +74,6 @@ class ArticleController < ApplicationController
 
   def show
     @article = Article.find(params[:id])
-  end
-
-  private
-  
-  def check_permissions
-    article = Article.find(params[:id])
-    post = article.post
-    if post.author_id!=current_user[:id] && current_user[:id] != 1
-      flash[:error] = "Недостаточно прав для данного действия!"
-      redirect_to post_url({:year => post.created_at.strftime('%Y'), :month => post.created_at.strftime('%m'), :day => post.created_at.strftime('%d'), :url => !post.url.nil? ? post.url : ''})
-    end
   end
 
 end
